Move database to app level
