+++ lib/home_page.dart
@@ -1,12 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:gcrdeviceconfigurator/data/database.dart';
 
 import 'home.dart';
 import 'settings_page.dart';
 
 class HomePage extends StatelessWidget {
+  final Database database;
+
   const HomePage({
     Key? key,
     required this.title,
+    required this.database,
   }) : super(key: key);
 
   final String title;
@@ -29,6 +33,6 @@ class HomePage extends StatelessWidget {
             ),
           ],
         ),
-        body: const Home());
+        body: Home(database: database));
   }
 }
